+++ CHANGELOG.md
@@ -7,6 +7,11 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- `buffers.focus_on_delete` config option to choose which buffer to focus after
+  a bufer is deleted.
+
 ## [0.2.0] - 2022-01-01
 
 ### Added